Added API Yaml files
Added the ability to return to login screen, as well as automatic returning upon incorrect login information. Removed sending data via Query string and moved it to using JSON request bodies. Added .yaml files for the 2 API's used in the system, one for use with the webpage, the other for external data entry.